requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
boto3==1.16.4
botocore==1.19.4
celery==4.4.7
celery-progress==0.0.14
cffi==1.14.3
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==2.9.2
Django==3.1.2
django-celery-results==1.2.1
django-cleanup==5.1.0
django-crispy-forms==1.9.2
django-storages==1.10.1
djangorestframework==3.12.1
flake8==3.8.4
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jwt==1.0.0
kombu==4.6.11
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.0.1
prompt-toolkit==3.0.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
urllib3==1.25.11
vine==1.3.0
wcwidth==0.2.5
Werkzeug==1.0.1
